Isolate the name of the 20 books with most data.
We have >19 hours per speaker for 20 speakers.

Next copy these top 20, after throwing out "the", to a file. Loop over
those globs in python, reading each mp3 in chunks of half-second,
sending them through mfcc and write each result in its own file, for
easy shuffling. Ensure that each mfcc is 49x13.

If we store the 3rd derivative though, that's
19 * 60 * 60 * 2 * 49*13*3 * 20 * 8 bytes, assuming 8 bytes per float of
the mfccs, which is 41.8 GB.

We should cap at 5 hours per speaker, at least at first, in that case.

Or we can store in raw form first, which is
19 * 60 * 60 * 2 * 20 * 4000 * 2 bytes = 21.8 GB because raw uses 2
bytes per sample, not 8 bytes.