Let a huge CNN model run overnight
Back of the envelope tells me this will take 24 minutes per epoch.

On smaller networks I'd been getting ~80-88% train accuracy, which is
kind of low. The dev accuracy was 5-10 % points behind that, which isn't
terrible.

I'm confused why the CNN with fewer parameters takes so much longer to
train than the FC.