94/90 when running this on the clean data
so we could probably get some mileage out of both increasing the size of
the network and the data set. test data is again really bad, at 67%
accuracy, which I don't understand, because we're not overfitting to the
dev set so I suspect something is wrong with the test set calculation.